+++ src/main/java/domain/result/GameResult.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.result;
 
 import domain.participant.Participant;
