+++ src/main/java/domain/result/Result.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.result;
 
 public enum Result {
     WIN(),
