+++ src/main/java/view/SuitMessage.java
@@ -1,6 +1,6 @@
 package view;
 
-import domain.Suit;
+import domain.card.Suit;
 
 import java.util.Collections;
 import java.util.Map;
