+++ src/test/java/domain/PlayerTest.java
@@ -1,5 +1,8 @@
 package domain;
 
+import domain.card.Card;
+import domain.card.Rank;
+import domain.card.Suit;
 import domain.participant.Participant;
 import domain.participant.Player;
 import org.junit.jupiter.api.BeforeEach;
